asset-register/requirements.txt
Flask>=3.0,<4
gunicorn>=22.0,<24
python-dotenv>=1.0,<2
